# gateway/openapi2-appengine.yaml
swagger: '2.0'
info:
  title: Polysnap API
  description: Sample API on API Gateway with an App Engine backend
  version: 1.0.0
schemes:
  - https
produces:
  - application/json
host: "polysnap-gateway-1i6xp2f783nz2.apigateway.cloud-final-402019.cloud.goog"
...
paths:
...
  /story/contact/{userId}:
    get:
      summary: Get contact stories of a user
      operationId: getContactStoryOfUser
      parameters:
        - name: userId
          in: path
          type: integer
          required: true
      x-google-backend:
        address: https://stories-dot-cloud-final-402019.oa.r.appspot.com
        path_translation: APPEND_PATH_TO_ADDRESS

      responses:
        '200':
          description: A successful response
          schema:
            type: string
...
  /chats/{id}:
    get:
      summary: Get a chat by ID
      operationId: findChat
      x-google-backend:
        address: https://messagesender-dot-cloud-final-402019.oa.r.appspot.com
        path_translation: APPEND_PATH_TO_ADDRESS
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        '200':
          description: A successful response

    delete:
      summary: Delete a chat by ID
      operationId: deleteChatById
      x-google-backend:
        address: https://messagesender-dot-cloud-final-402019.oa.r.appspot.com
        path_translation: APPEND_PATH_TO_ADDRESS
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        '200':
          description: A successful response
...